[OE-2034] fixed transport pagination not displaying all the data. this removes the horrible multi-table union query and replaces it with one join. dupes are also removed.

+++ protected/controllers/TransportController.php
@@ -88,116 +88,101 @@ class TransportController extends BaseController
 
 		if ($from && $to) {
 			$wheresql1 = " and session.date >= '$from' and session.date <= '$to' ";
-			$wheresql2 = " and cancelled_booking.date >= '$from' and cancelled_booking.date <= '$to' ";
+			$wheresql2 = " and cb.date >= '$from' and cb.date <= '$to' ";
 		} else {
 			$wheresql1 = $wheresql2 = null;
 		}
 
-		$sql = '';
-
-		if ($include_bookings) {
-			$sql = "select element_operation.id as eoid, booking.id as checkid, patient.id as pid, event.id as evid, contact.first_name, contact.last_name, patient.hos_num, site.short_name as location, element_operation.eye_id, firm.pas_code as firm, element_operation.decision_date, pr.name as priority, subspecialty.ref_spec as subspecialty, session.date as session_date, session.start_time as session_time, element_operation.status, 'Booked' as method, transport_list.id as transport, booking.created_date as order_date, ward.name as ward_name from booking
-				join session on booking.session_id = session.id
-				join theatre on session.theatre_id = theatre.id
-				join site on theatre.site_id = site.id
-				join element_operation on element_operation.id = booking.element_operation_id
-				join priority pr on pr.id = element_operation.priority_id
-				join event on element_operation.event_id = event.id
-				join episode on event.episode_id = episode.id
-				join firm on episode.firm_id = firm.id
-				join service_subspecialty_assignment on firm.service_subspecialty_assignment_id = service_subspecialty_assignment.id
-				join subspecialty on service_subspecialty_assignment.subspecialty_id = subspecialty.id
-				join patient on episode.patient_id = patient.id
-				join contact on contact.parent_id = patient.id and contact.parent_class = 'Patient'
-				left join transport_list on (transport_list.item_table = 'booking' and transport_list.item_id = booking.id)
-				join ward on booking.ward_id = ward.id
-				where session.date >= '$today' $wheresql1 and element_operation.status != 3
-				and site.id not in (3,5)
-				and (transport_list.id is null or substr(transport_list.last_modified_date,1,10) = '$today')";
-		}
-
-		if ($include_reschedules) {
-			if ($sql) $sql .= " UNION ";
-			$sql .= "select element_operation.id as eoid, booking.id as checkid, patient.id as pid, event.id as evid, contact.first_name, contact.last_name, patient.hos_num, site.short_name as location, element_operation.eye_id, firm.pas_code as firm, element_operation.decision_date, pr.name as priority, subspecialty.ref_spec as subspecialty, session.date as session_date, session.start_time as session_time, element_operation.status, 'Rescheduled' as method, transport_list.id as transport, cancelled_booking.created_date as order_date, ward.name as ward_name from booking
-				join session on booking.session_id = session.id
-				join cancelled_booking on cancelled_booking.element_operation_id = booking.element_operation_id
-				join theatre on session.theatre_id = theatre.id
-				join site on theatre.site_id = site.id
-				join element_operation on element_operation.id = booking.element_operation_id
-				join priority pr on pr.id = element_operation.priority_id
-				join event on element_operation.event_id = event.id
-				join episode on event.episode_id = episode.id
-				join firm on episode.firm_id = firm.id
-				join service_subspecialty_assignment on firm.service_subspecialty_assignment_id = service_subspecialty_assignment.id
-				join subspecialty on service_subspecialty_assignment.subspecialty_id = subspecialty.id
-				join patient on episode.patient_id = patient.id
-				join contact on contact.parent_id = patient.id and contact.parent_class = 'Patient'
-				left join transport_list on (transport_list.item_table = 'booking' and transport_list.item_id = booking.id)
-				join ward on booking.ward_id = ward.id
-				where session.date >= '$today' $wheresql1 and element_operation.status = 3
-				and site.id not in (3,5)
-				and (transport_list.id is null or substr(transport_list.last_modified_date,1,10) = '$today')";
-		}
-
-		if ($include_cancellations) {
-			if ($sql) $sql .= " UNION ";
-			$sql .= "select element_operation.id as eoid, cancelled_booking.id as checkid, patient.id as pid, event.id as evid, contact.first_name, contact.last_name, patient.hos_num, site.short_name as location, element_operation.eye_id, firm.pas_code as firm, element_operation.decision_date, pr.name as priority, subspecialty.ref_spec as subspecialty, cancelled_booking.date as session_date, cancelled_booking.start_time as session_time, element_operation.status, 'Cancelled' as method, transport_list.id as transport, cancelled_booking.created_date as order_date, 'Unknown' as ward_name from cancelled_booking
-				join theatre on cancelled_booking.theatre_id = theatre.id
-				join site on theatre.site_id = site.id
-				join element_operation on element_operation.id = cancelled_booking.element_operation_id
-				join priority pr on pr.id = element_operation.priority_id
-				join event on element_operation.event_id = event.id
-				join episode on event.episode_id = episode.id
-				join firm on episode.firm_id = firm.id
-				join service_subspecialty_assignment on firm.service_subspecialty_assignment_id = service_subspecialty_assignment.id
-				join subspecialty on service_subspecialty_assignment.subspecialty_id = subspecialty.id
-				join patient on episode.patient_id = patient.id
-				join contact on contact.parent_id = patient.id and contact.parent_class = 'Patient'
-				left join transport_list on (transport_list.item_table = 'cancelled_booking' and transport_list.item_id = cancelled_booking.id)
-				where cancelled_booking.date >= '$today' $wheresql2 and element_operation.status != 3
-				and site.id not in (3,5)
-				and (transport_list.id is null or substr(transport_list.last_modified_date,1,10) = '$today')";
-		}
-
 		$offset = ($this->items_per_page * ($this->page-1));
-
-		$sql .= " ORDER BY session_date asc, session_time asc, order_date desc";
 
 		$data = array();
 		$data_all = array();
-		$done = array();
 
-		foreach (Yii::app()->db->createCommand($sql)->query() as $i => $row) {
-			if (!in_array($this->getDisplayArray($row),$done)) {
-				if ($i >= $offset && count($data) < $this->items_per_page) {
-					$data[] = $row;
-				}
-				$data_all[] = $row;
-				$done[] = $this->getDisplayArray($row);
+		$exclude_sites = array(3,5);
+		$exclude_theatres = array();
+
+		// cache theatre->site associations for page load speed
+		$sites = array();
+		foreach (Yii::app()->db->createCommand()
+			->select('theatre.id, site.name, site.short_name, site.id as site_id')
+			->from('theatre')
+			->join('site','theatre.site_id = site.id')
+			->queryAll() as $row) {
+			if (in_array($row['site_id'],$exclude_sites)) {
+				$exclude_theatres[] = $row['id'];
+			} else {
+				$sites[$row['id']] = $row['short_name'] ? $row['short_name'] : $row['name'];
 			}
 		}
+
+		if (!empty($exclude_theatres)) {
+			$wheresql3 = ' and session.theatre_id not in ('.implode(',',$exclude_theatres).') ';
+			$wheresql4 = ' and cb.theatre_id not in ('.implode(',',$exclude_theatres).') ';
+		} else {
+			$wheresql3 = $wheresql4 = '';
+		}
+
+		foreach (Yii::app()->db->createCommand()
+			->selectDistinct('element_operation.id as eoid, element_operation.priority_id, booking.id as booking_id, patient.id as pid, event.id as evid, contact.first_name,
+				contact.last_name, patient.hos_num, element_operation.eye_id, firm.pas_code as firm, element_operation.decision_date, subspecialty.ref_spec as subspecialty,
+				session.date as session_date, session.start_time as session_time, element_operation.status, transport_list.id as transport,
+				coalesce(booking.created_date,cb.created_date) as order_created_date, ward.name as ward_name, cb.id as cancelled_booking_id, cb.date as cancelled_session_date,
+				session.theatre_id, coalesce(session.date,cb.date) as order_date, cb.start_time as cancelled_session_time, session.id as session_id, cb.theatre_id as theatre_id2,
+				coalesce(session.start_time,cb.start_time) as order_time, cb.date as cancelled_booking_date')
+			->from('element_operation')
+			->leftJoin('booking','booking.element_operation_id = element_operation.id')
+			->leftJoin('session',"booking.session_id = session.id and session.date >= '$today'")
+			->join('event','element_operation.event_id = event.id')
+			->join('episode','event.episode_id = episode.id')
+			->join('firm','episode.firm_id = firm.id')
+			->join('service_subspecialty_assignment','firm.service_subspecialty_assignment_id = service_subspecialty_assignment.id')
+			->join('subspecialty','service_subspecialty_assignment.subspecialty_id = subspecialty.id')
+			->join('patient','episode.patient_id = patient.id')
+			->join('contact',"contact.parent_id = patient.id and contact.parent_class = 'Patient'")
+			->leftJoin('transport_list',"transport_list.item_table = 'booking' and transport_list.item_id = booking.id and substr(transport_list.last_modified_date,1,10) = '$today'")
+			->leftJoin('ward','booking.ward_id = ward.id')
+			->leftJoin('(select *,max(id) as maxid from cancelled_booking group by element_operation_id) cb',"element_operation.id = cb.element_operation_id and cb.date >= '$today'")
+			->leftJoin('transport_list transport_list2',"transport_list2.item_table = 'cancelled_booking' and transport_list2.item_id = cb.id and substr(transport_list2.last_modified_date,1,10) = '$today'")
+			->where("(booking.id is not null $wheresql1 $wheresql3) or (booking.id is null and cb.id is not null $wheresql2 $wheresql4 )")
+			->order("order_date asc, order_time asc, order_created_date desc")
+			->queryAll() as $i => $row) {
+
+			if ($row['booking_id']) {
+				if ($row['status'] == 3) {
+					$row['method'] = 'Rescheduled';
+				} else {
+					$row['method'] = 'Booked';
+				}
+				$ts = strtotime($row['session_date'].' '.$row['session_time']);
+				$row['checkid'] = $row['booking_id'];
+				$row['location'] = $sites[$row['theatre_id']];
+			} else {
+				$row['method'] = 'Cancelled';
+				$ts = strtotime($row['cancelled_session_date'].' '.$row['cancelled_session_time']);
+				$row['checkid'] = $row['cancelled_booking_id'];
+				$row['session_date'] = $row['cancelled_session_date'];
+				$row['session_time'] = $row['cancelled_session_time'];
+				$row['ward_name'] = 'Unknown';
+				$row['location'] = $sites[$row['theatre_id2']];
+			}
+
+			if (($include_bookings && $row['method'] == 'Booked') || ($include_reschedules && $row['method'] == 'Rescheduled') || ($include_cancellations && $row['method'] == 'Cancelled')) {
+				if (count($data_all) >= $offset && count($data) < $this->items_per_page) {
+					while (isset($data[$ts])) $ts++;
+					$data[$ts] = $row;
+				}
+				while (isset($data_all[$ts])) $ts++;
+				$data_all[$ts] = $row;
+			}
+		}
+
+		ksort($data);
+		ksort($data_all);
 
 		$this->total_items = count($data_all);
 		$this->pages = ceil($this->total_items / $this->items_per_page);
 
 		return array('bookings' => $data, 'bookings_all' => $data_all);
-	}
-
-	public function getDisplayArray($row) {
-		return array(
-			'hos_num' => $row['hos_num'],
-			'first_name' => $row['first_name'],
-			'last_name' => $row['last_name'],
-			'tci_date' => $row['session_date'],
-			'admission_time' => $row['session_time'],
-			'location' => $row['location'],
-			'ward_name' => $row['ward_name'],
-			'method' => $row['method'],
-			'firm' => $row['firm'],
-			'subspecialty' => $row['subspecialty'],
-			'decision_date' => $row['decision_date'],
-			'priority' => $row['priority'],
-		);
 	}
 
 	public function actionDigest() {
@@ -296,22 +281,26 @@ class TransportController extends BaseController
 
 		if (isset($_REQUEST['cancelled']) && is_array($_REQUEST['cancelled'])) {
 			foreach ($_REQUEST['cancelled'] as $cancelled_booking_id) {
-				$c = TransportList::Model()->find('item_table = ? and item_id = ? and status = ?',array('cancelled_booking',$cancelled_booking_id,1));
+				if ($cancelled_booking = CancelledBooking::model()->findByPk($cancelled_booking_id)) {
+					foreach (CancelledBooking::model()->findAll('element_operation_id=?',array($cancelled_booking->element_operation_id)) as $cb) {
+						$c = TransportList::Model()->find('item_table = ? and item_id = ? and status = ?',array('cancelled_booking',$cb->id,1));
 
-				if (!$c) {
-					$c = new TransportList;
-					$c->item_table = 'cancelled_booking';
-					$c->item_id = $cancelled_booking_id;
-					$c->status = 1;
-					if (!$c->save()) {
-						throw new SystemException('Unable to save transport_list item: '.print_r($c->getErrors(),true));
+						if (!$c) {
+							$c = new TransportList;
+							$c->item_table = 'cancelled_booking';
+							$c->item_id = $cb->id;
+							$c->status = 1;
+							if (!$c->save()) {
+								throw new SystemException('Unable to save transport_list item: '.print_r($c->getErrors(),true));
+							}
+						} else {
+							/*
+							if (!$c->delete()) {
+								throw new SystemException('Unable to delete transport_list item: '.print_r($c->getErrors(),true));
+							}
+							*/
+						}
 					}
-				} else {
-					/*
-					if (!$c->delete()) {
-						throw new SystemException('Unable to delete transport_list item: '.print_r($c->getErrors(),true));
-					}
-					*/
 				}
 			}
 		}
@@ -330,7 +319,7 @@ class TransportController extends BaseController
 		$data = $this->getBookings();
 
 		foreach ($data['bookings_all'] as $row) {
-			echo '"'.$row['hos_num'].'","'.trim($row['first_name']).'","'.trim($row['last_name']).'","'.$row['session_date'].'","'.$row['session_time'].'","'.$row['location'].'","'.$row['ward_name'].'","'.$row['method'].'","'.$row['firm'].'","'.$row['subspecialty'].'","'.$row['decision_date'].'","'.$row['priority'].'"'."\n";
+			echo '"'.$row['hos_num'].'","'.trim($row['first_name']).'","'.trim($row['last_name']).'","'.$row['order_date'].'","'.$row['order_time'].'","'.$row['location'].'","'.$row['ward_name'].'","'.$row['method'].'","'.$row['firm'].'","'.$row['subspecialty'].'","'.$row['decision_date'].'","'.($row['priority_id'] == 1 ? 'Routine' : 'Urgent').'"'."\n";
 		}
 	}
